Initial re-organisation and re-factoring of Highlighting feature
