+++ src/content-tooltip/interactions.ts
@@ -1,11 +1,7 @@
 import { browser } from 'webextension-polyfill-ts'
 
 import { delayed, getPositionState, getTooltipState } from './utils'
-import {
-    createAndCopyDirectLink,
-    createAnnotation,
-    createHighlight,
-} from '../direct-linking/content_script/interactions'
+import { createAndCopyDirectLink } from '../direct-linking/content_script/interactions'
 import { setupUIContainer, destroyUIContainer } from './components'
 import {
     remoteFunction,
@@ -14,6 +10,10 @@ import {
 import { injectCSS } from '../search-injection/dom'
 import { conditionallyShowHighlightNotification } from './onboarding-interactions'
 import { TooltipInteractionInterface } from 'src/content-tooltip/types'
+import {
+    createAnnotation,
+    renderHighlightAndCreateAnnotation,
+} from 'src/highlighting/ui'
 
 const openOptionsRPC = remoteFunction('openOptionsTab')
 let mouseupListener = null
@@ -76,7 +76,7 @@ export const insertTooltip = async ({ toolbarNotifications }) => {
     showTooltip = await setupUIContainer(target, {
         createAndCopyDirectLink,
         createAnnotation,
-        createHighlight,
+        createHighlight: renderHighlightAndCreateAnnotation,
         openSettings: () => openOptionsRPC('settings'),
         destroyTooltip: async () => {
             manualOverride = true
