+++ src/highlighting/README.md
@@ -1,0 +1,5 @@
+# Highlights
+
+Highlights are selections across text or other DOM nodes on a page.
+
+A highlight is a subset of an Annotation (An Annotation is a Highlight without a Note)
