+++ src/highlighting/ui/anchoring/anchoring/LICENSE
