+++ src/highlighting/ui/anchoring/anchoring/html.coffee
