+++ src/highlighting/ui/anchoring/anchoring/pdf.coffee
