+++ src/highlighting/ui/anchoring/anchoring/range.coffee
