+++ src/highlighting/ui/anchoring/anchoring/types.coffee
