+++ src/highlighting/ui/anchoring/anchoring/util.coffee
