+++ src/highlighting/ui/anchoring/anchoring/xpath.coffee
