+++ src/highlighting/ui/highlight-interactions.ts
@@ -1,21 +1,93 @@
-import { highlightAnnotation } from 'src/direct-linking/content_script/rendering'
-import { getOffsetTop } from '../utils'
+import { getOffsetTop } from 'src/sidebar-overlay/utils'
 import { Annotation } from 'src/sidebar-overlay/sidebar/types'
+import { Highlight } from 'src/highlighting/types'
+import { retryUntil } from 'src/util/retry-until'
+import { descriptorToRange, markRange } from './anchoring/index'
+import * as AllRaven from 'raven-js'
+const styles = require('src/highlighting/ui/styles.css')
+const Raven = AllRaven['default']
 
-const styles = require('src/direct-linking/content_script/styles.css')
+/**
+ * Given an array of highlight objects, highlights all of them.
+ */
+export const renderHighlights = async (
+    highlights: Highlight[],
+    openSidebar: (args: { activeUrl?: string }) => void,
+    focusOnHighlight?: (url: string) => void,
+    hoverHighlightContainer?: (url: string) => void,
+) => {
+    await Promise.all(
+        highlights.map(async highlight =>
+            renderHighlight(
+                highlight,
+                focusOnHighlight,
+                hoverHighlightContainer,
+                openSidebar,
+            ),
+        ),
+    )
+}
+
+export const renderHighlight = async (
+    highlight: Highlight,
+    focusOnAnnotation,
+    hoverAnnotationContainer,
+    openSidebar,
+) => {
+    const baseClass = styles['memex-highlight']
+    try {
+        await Raven.context(async () => {
+            const descriptor = highlight.anchors
+                ? highlight.anchors[0].descriptor
+                : highlight.selector.descriptor
+
+            Raven.captureBreadcrumb({
+                message: 'annotation-selector-received',
+                category: 'annotations',
+                data: highlight,
+            })
+
+            const range = await retryUntil(
+                () => descriptorToRange({ descriptor }),
+                _range => _range !== null,
+                {
+                    intervalMiliseconds: 200,
+                    timeoutMiliseconds: 5000,
+                },
+            )
+
+            markRange({ range, cssClass: baseClass })
+
+            attachEventListenersToNewHighlights(
+                highlight,
+                focusOnAnnotation,
+                hoverAnnotationContainer,
+                openSidebar,
+            )
+        })
+    } catch (e) {
+        console.error(
+            'MEMEX: Error during annotation anchoring/highlighting:',
+            e,
+        )
+        console.error(e.stack)
+        return false
+    }
+
+    return true
+}
 
 /**
  * Scrolls to the highlight of the given annotation on the current page.
  */
-export const scrollToHighlight = ({ url }: Annotation) => {
+export const scrollToHighlight = ({ url }: Highlight) => {
     const baseClass = styles['memex-highlight']
     const $highlight = document.querySelector(
         `.${baseClass}[data-annotation="${url}"]`,
-    )
+    ) as HTMLElement
 
     if ($highlight) {
-        const top =
-            getOffsetTop($highlight as HTMLElement) - window.innerHeight / 2
+        const top = getOffsetTop($highlight) - window.innerHeight / 2
         window.scrollTo({ top, behavior: 'smooth' })
         // The pixels scrolled need to be returned in order to restrict
         // scrolling when mouse is over the sidebar.
@@ -24,7 +96,6 @@ export const scrollToHighlight = ({ url }: Annotation) => {
         console.error('MEMEX: Oops, no highlight found to scroll to')
     }
 }
-
 /**
  * Given an annotation object, highlights that text and removes other highlights
  * from the page.
@@ -37,66 +108,46 @@ export const highlightAndScroll = (annotation: Annotation) => {
 }
 
 /**
- * Given an array of annotation objects, highlights all of them.
- */
-export const highlightAnnotations = async (
-    annotations: Annotation[],
-    openSidebar: (args: { activeUrl?: string }) => void,
-    focusOnAnnotation?: (url: string) => void,
-    hoverAnnotationContainer?: (url: string) => void,
-) => {
-    await Promise.all(
-        annotations.map(async annotation =>
-            highlightAnnotation(
-                { annotation },
-                openSidebar,
-                focusOnAnnotation,
-                hoverAnnotationContainer,
-            ),
-        ),
-    )
-}
-
-/**
  * Attaches event listeners to the highlights for hovering/focusing on the
  * annotation in sidebar.
- * @param {Annotation} annotation The annotation to which the listeners are going to be attached
+ * @param {Highlight} highlight The annotation to which the listeners are going to be attached
  * @param {function} focusOnAnnotation Function when called will set the sidebar container to active state
  * @param {function} hoverAnnotationContainer Function when called will set the sidebar container to hover state
+ * @param openSidebar
  */
 export const attachEventListenersToNewHighlights = (
-    annotation: Annotation,
-    focusOnAnnotation: (url: string) => void,
-    hoverAnnotationContainer: (url: string) => void,
+    highlight: Highlight,
+    focusOnAnnotation: (url: string) => void = _ => undefined,
+    hoverAnnotationContainer: (url: string) => void = _ => undefined,
     openSidebar: (args: { activeUrl?: string }) => void,
 ) => {
     const newHighlights = document.querySelectorAll(
         `.${styles['memex-highlight']}:not([data-annotation])`,
     )
-    newHighlights.forEach(highlight => {
-        ;(highlight as HTMLElement).dataset.annotation = annotation.url
+    newHighlights.forEach(highlightEl => {
+        ;(highlightEl as HTMLElement).dataset.annotation = highlight.url
 
         const clickListener = async e => {
             e.preventDefault()
             if (!e.target.dataset.annotation) {
                 return
             }
-            openSidebar({ activeUrl: annotation.url })
+            openSidebar({ activeUrl: highlight.url })
             removeHighlights(true)
-            makeHighlightDark(annotation)
-            focusOnAnnotation(annotation.url)
+            makeHighlightDark(highlight)
+            focusOnAnnotation(highlight.url)
         }
-        highlight.addEventListener('click', clickListener, false)
+        highlightEl.addEventListener('click', clickListener, false)
 
         const mouseenterListener = e => {
             if (!e.target.dataset.annotation) {
                 return
             }
             removeMediumHighlights()
-            makeHighlightMedium(annotation)
-            hoverAnnotationContainer(annotation.url)
+            makeHighlightMedium(highlight)
+            hoverAnnotationContainer(highlight.url)
         }
-        highlight.addEventListener('mouseenter', mouseenterListener, false)
+        highlightEl.addEventListener('mouseenter', mouseenterListener, false)
 
         const mouseleaveListener = e => {
             if (!e.target.dataset.annotation) {
@@ -105,10 +156,9 @@ export const attachEventListenersToNewHighlights = (
             removeMediumHighlights()
             hoverAnnotationContainer(null)
         }
-        highlight.addEventListener('mouseleave', mouseleaveListener, false)
+        highlightEl.addEventListener('mouseleave', mouseleaveListener, false)
     })
 }
-
 /**
  * Removes the medium class from all the highlights making them light.
  */
@@ -121,12 +171,11 @@ export const removeMediumHighlights = () => {
     )
     prevHighlights.forEach(highlight => highlight.classList.remove(mediumClass))
 }
-
 /**
  * Makes the given annotation as a medium highlight.
  * @param {string} url PK of the annotation to make medium
  */
-export const makeHighlightMedium = ({ url }: Annotation) => {
+export const makeHighlightMedium = ({ url }: Highlight) => {
     // Make the current annotation as a "medium" highlight.
     const baseClass = styles['memex-highlight']
     const mediumClass = styles['medium']
@@ -135,11 +184,10 @@ export const makeHighlightMedium = ({ url }: Annotation) => {
     )
     highlights.forEach(highlight => highlight.classList.add(mediumClass))
 }
-
 /**
  * Makes the highlight a dark highlight.
  */
-export const makeHighlightDark = ({ url }: Annotation) => {
+export const makeHighlightDark = ({ url }: Highlight) => {
     const baseClass = styles['memex-highlight']
     const highlights = document.querySelectorAll(
         `.${baseClass}[data-annotation="${url}"]`,
@@ -149,7 +197,6 @@ export const makeHighlightDark = ({ url }: Annotation) => {
         highlight.classList.add(styles['dark'])
     })
 }
-
 /**
  * Removes all highlight elements in the current page.
  * If `onlyRemoveDarkHighlights` is true, only dark highlights will be removed.
@@ -168,7 +215,6 @@ export const removeHighlights = (onlyRemoveDarkHighlights = false) => {
         highlights.forEach(highlight => _removeHighlight(highlight))
     }
 }
-
 /**
  * Finds each annotation's position in page, sorts it by the position and
  * returns the sorted annotations.
@@ -176,9 +222,7 @@ export const removeHighlights = (onlyRemoveDarkHighlights = false) => {
 export const sortAnnotationsByPosition = (annotations: Annotation[]) => {
     const offsetTopObjects = annotations.map((annotation, index) => {
         const firstHighlight = document.querySelector(
-            `.${styles['memex-highlight']}[data-annotation="${
-                annotation.url
-            }"]`,
+            `.${styles['memex-highlight']}[data-annotation="${annotation.url}"]`,
         )
         return {
             index,
