+++ src/highlighting/ui/styles.css
