+++ src/sidebar-overlay/comment-box/reducer.ts
@@ -2,7 +2,7 @@ import { createReducer } from 'redux-act'
 
 import * as actions from './actions'
 import State from './types'
-import { Anchor } from 'src/direct-linking/content_script/interactions'
+import { Anchor } from 'src/highlighting/types'
 
 export const defaultState: State = {
     showCommentBox: false,
