+++ src/sidebar-overlay/content_script/ribbon-interactions.ts
@@ -9,7 +9,7 @@ import { setupRibbonAndSidebarUI, destroyRibbonAndSidebarUI } from '..'
 import { getSidebarState } from '../utils'
 import { getTooltipState } from 'src/content-tooltip/utils'
 import { createRootElement, destroyRootElement } from './rendering'
-import { removeHighlights } from './highlight-interactions'
+import { removeHighlights } from '../../highlighting/ui/highlight-interactions'
 import AnnotationsManager from 'src/sidebar-overlay/annotations-manager'
 import ToolbarNotifications from 'src/toolbar-notification/content_script'
 import { insertTooltip, removeTooltip } from 'src/content-tooltip/interactions'
